refactor: move validateIdentity to TopUpRepository
