+++ wallet/src/de/schildbach/wallet/ui/dashpay/PlatformRepo.kt
@@ -33,7 +33,6 @@ import de.schildbach.wallet.database.entity.BlockchainIdentityBaseData
 import de.schildbach.wallet.database.entity.BlockchainIdentityData
 import de.schildbach.wallet.database.entity.DashPayContactRequest
 import de.schildbach.wallet.database.entity.DashPayProfile
-import de.schildbach.wallet.database.entity.Invitation
 import de.schildbach.wallet.database.entity.BlockchainIdentityConfig
 import de.schildbach.wallet.livedata.Resource
 import de.schildbach.wallet.livedata.SeriousError
@@ -1046,71 +1045,6 @@ class PlatformRepo @Inject constructor(
             })
             walletApplication.broadcastTransaction(cftx)
         }
-    }
-
-    /**
-     * validates an invite
-     *
-     * @return Returns true if it is valid, false if the invite has been used.
-     *
-     * @throws Exception if the invite is invalid
-     */
-
-    fun validateInvitation(invite: InvitationLinkData): Boolean {
-        val stopWatch = Stopwatch.createStarted()
-        var tx = platform.client.getTransaction(invite.cftx)
-        log.info("validateInvitation: obtaining transaction info took $stopWatch")
-        //TODO: remove when iOS uses big endian
-        if (tx == null) {
-            tx =
-                platform.client.getTransaction(Sha256Hash.wrap(invite.cftx).reversedBytes.toHex())
-        }
-        if (tx != null) {
-            val cfTx = AssetLockTransaction(Constants.NETWORK_PARAMETERS, tx)
-            val identity = platform.identities.get(cfTx.identityId.toStringBase58())
-            if (identity == null) {
-                // determine if the invite has enough credits
-                if (cfTx.lockedOutput.value < Constants.DASH_PAY_INVITE_MIN) {
-                    val reason = "Invite does not have enough credits ${cfTx.lockedOutput.value} < ${Constants.DASH_PAY_INVITE_MIN}"
-                    log.warn(reason)
-                    log.info("validateInvitation took $stopWatch")
-                    throw InsufficientMoneyException(cfTx.lockedOutput.value, reason)
-                }
-                return try {
-                    DumpedPrivateKey.fromBase58(Constants.NETWORK_PARAMETERS, invite.privateKey)
-                    // TODO: when all instantsend locks are deterministic, we don't need the catch block
-                    try {
-                        InstantSendLock(
-                            Constants.NETWORK_PARAMETERS,
-                            Utils.HEX.decode(invite.instantSendLock),
-                            InstantSendLock.ISDLOCK_VERSION
-                        )
-                    } catch (e: Exception) {
-                        InstantSendLock(
-                            Constants.NETWORK_PARAMETERS,
-                            Utils.HEX.decode(invite.instantSendLock),
-                            InstantSendLock.ISLOCK_VERSION
-                        )
-                    }
-                    log.info("Invite is valid and took $stopWatch")
-                    true
-                } catch (e: AddressFormatException.WrongNetwork) {
-                    log.warn("Invite has private key from wrong network: $e and took $stopWatch")
-                    throw e
-                } catch (e: AddressFormatException) {
-                    log.warn("Invite has invalid private key: $e and took $stopWatch")
-                    throw e
-                } catch (e: Exception) {
-                    log.warn("Invite has invalid instantSendLock: $e and took $stopWatch")
-                    throw e
-                }
-            } else {
-                log.warn("Invitation has been used: ${identity.id} and took $stopWatch")
-                return false
-            }
-        }
-        log.warn("Invitation uses an invalid transaction ${invite.cftx} and took $stopWatch")
-        throw IllegalArgumentException("Invitation uses an invalid transaction ${invite.cftx}")
     }
 
     fun clearBlockchainIdentityData() {
